requirements.txt
fastapi
sqlalchemy
uvicorn
passlib[bcrypt]
PyJWT
python-multipart